+++ CHANGES.md
@@ -3,6 +3,9 @@ Change Log
 
 Beta Releases
 -------------
+### b22 - 2013-11-101
+* Added `PolylineVolumeGeometry` and `PolylineVolumeGeometryOutline`.
+* Added `Shapes.compute2DCircle`.
 
 ### b21 - 2013-10-01
 
@@ -66,9 +69,6 @@ Beta Releases
 * Fixed geometries not closing completely. [#1093](https://github.com/AnalyticalGraphicsInc/cesium/issues/1093)
 * Fixed `EllipsoidTangentPlane.projectPointOntoPlane` for tangent planes on an ellipsoid other than the unit sphere.
 * `CompositePrimitive.add` now returns the added primitive. This allows us to write more concise code.
-* Added `PolylineVolumeGeometry` and `PolylineVolumeGeometryOutline`.
-* Added `Shapes.compute2DCircle`.
-
         var p = new Primitive(/* ... */);
         primitives.add(p);
         return p;